requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
text-unidecode==1.3
uritemplate==4.1.1
